Move duplicated Latex parser to shared module
